siteDbRestore ok

+++ bash/ubuntu-16.x/framework/dump.sh
@@ -2,10 +2,10 @@
 
 frameworkDumpArgs() {
   _ARGUMENTS=(
-    [0]='site_dir s "Root directory of site" true'
-    [1]='dump_dir d "Target directory for dumps" true'
-    [2]='prefix px "Prefix for final file" false'
-    [3]='zip zip "Use ZIP" false'
+    [0]='dump_dir d "Target directory for dumps" true'
+    [1]='prefix px "Prefix for final file" false'
+    [2]='zip zip "Use ZIP" false'
+    [3]='zip_only zo "Remove original file after ZIP it" false'
     [4]='host H "Database host server" false'
     [5]='port P "Database host port" false'
     [6]='database db "Database name" false'
@@ -47,16 +47,24 @@ frameworkDump() {
   fi;
 
   # Build dump name.
-  DUMP_FULL_PATH=${DUMP_DIR}"/"${PREFIX}${DATABASE}"-"$(wex date/timeFileName)".sql"
+  DUMP_FILE_NAME=${PREFIX}${DATABASE}"-"$(wex date/timeFileName)".sql"
+  DUMP_FULL_PATH=${DUMP_DIR}"/"${DUMP_FILE_NAME}
 
   mysqldump -h${HOST} -P${PORT} -u${USER} ${PASSWORD} ${DATABASE} > ${DUMP_FULL_PATH}
 
   if [[ ${ZIP} == true ]]; then
-     zip ${DUMP_FULL_PATH}".zip" ${DUMP_FULL_PATH} -q
-     echo ${DUMP_FULL_PATH}".zip"
-     rm -rf ${DUMP_FULL_PATH}
-     return
+    zip ${DUMP_FULL_PATH}".zip" ${DUMP_FULL_PATH} -q -j
+
+    # Expect only zipped file
+    # Remove original.
+    if [[ ! -z "${ZIP_ONLY+x}" ]]; then
+      rm -rf ${DUMP_FULL_PATH}
+      # Return the zipped file name.
+      echo ${DUMP_FULL_PATH}".zip"
+      return
+    fi;
   fi
 
+  # Return dump file name.
   echo ${DUMP_FULL_PATH}
 }
